Refactor metadata conversion into its own module
